Use plain i64 for mempool usage
It seems there is no point to use atomic var for memory usage in mempool
- in the C++ version only the "NoBarrier" methods are called from the
AtomicPointer class. Therefore, this changes the type to be plain i64.

Also some changes on the atomic_cell: instead of using Cell to implement
this, now we just use *mut T, following the same pattern as in
std::sync::atomic. Also added a new type AtomicI64 which implements
atomic i64 with compiler fence. Naturally, atomic_cell is renamed to
atomic to avoid confusion.