Rename MemPool to Arena
This renames the MemPool struct to Arena (which is the same as the c++
impl...). Also, the Arena APIs now return *mut u8 instead of &mut [u8].
This offers more flexibility, and more importantly, to avoid
"lifetime chaining effect" on the Arena. However, this also means that
clients using the Arena struct needs to make sure the pointers won't be
sticking around after Arena goes out of scope.